+++ client/src/features/myCollections/SelectionList.tsx
@@ -1,12 +1,8 @@
 import React from 'react'
 
-import SelectionOption from './SelectionOption'
-
 const SelectionList: React.FC<{
-  listOfRecords: Array<string>
-  selected: Array<string>
-  handleSelection: (value: Array<string> | string) => void
-}> = ({ listOfRecords, handleSelection, selected }) => (
+  children: Array<JSX.Element>
+}> = ({ children }) => (
   <div
     className="rounded-2 border"
     style={{
@@ -15,13 +11,7 @@ const SelectionList: React.FC<{
     }}
     data-testid="user-collections-list-container"
   >
-    {listOfRecords.map((record) => (
-      <SelectionOption
-        record={record}
-        selectedRecords={selected}
-        handleSelection={handleSelection}
-      />
-    ))}
+    {children}
   </div>
 )